# docs/specs/openapi.yaml
openapi: 3.0.0
info:
  title: Memes Concept API
  version: 1.0.0
  description: API for the Memes Concept MVP store.
servers:
  - url: http://localhost:3001
    description: Local Development Server

paths:
  /products:
    get:
      summary: List all products
      operationId: listProducts
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create a product (Admin)
      operationId: createProduct
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                images:
                  type: array
                  items:
                     type: string
              required:
                - name
                - price
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized

  /products/{id}:
    get:
      summary: Get a product by ID or Slug
      operationId: getProduct
      parameters:
        - name: id
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
    put:
      summary: Update a product (Admin)
      operationId: updateProduct
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated
    delete:
      summary: Delete a product (Admin)
      operationId: deleteProduct
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted

  /admin/products/upload:
    post:
      summary: Upload an image to Cloudinary (Base64)
      operationId: uploadImage
      security:
        - AdminAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: Base64 encoded image string
      responses:
        '200':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '500':
          description: Upload failed

  /checkout:
    post:
      summary: Create a Stripe Checkout Session
      operationId: createCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutSessionRequest'
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
        '400':
          description: Invalid input or stock issues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /webhook/stripe:
    post:
      summary: Handle Stripe Webhooks
      description: Receives events from Stripe (e.g., payment_intent.succeeded)
      responses:
        '200':
          description: Webhook received

components:
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-key
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier (MongoDB _id)
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        images:
          type: array
          items:
            type: string
          description: Array of image URLs
        slug:
          type: string
          description: URL-friendly identifier
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'
      required:
        - id
        - name
        - price
        - slug
        - images

    ProductVariant:
      type: object
      properties:
        size:
          type: string
          enum: [S, M, L, XL, XXL]
        color:
          type: string
          nullable: true
        stock:
          type: integer


    CheckoutSessionRequest:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
        variant:
          type: object
          properties:
            size:
              type: string
            color:
              type: string
      required:
        - productId
        - quantity

    CheckoutSessionResponse:
      type: object
      properties:
        clientSecret:
          type: string
          description: Stripe Embeded Checkout Client Secret

    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string